Update README.md
I was not able to embed the Gif I created because Imgur kept converting it to .mp4 even though it is a .gif on my laptop. I could not solve this problem therefore instead I have put a hyperlink that will redirect you to the imgur site with my walkthrough. Here is the link again: https://imgur.com/gallery/kovDJsO and this is a direct link to the video   https://i.imgur.com/mPK9NYg.mp4